requirements.txt
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.24.0
PyOpenGL>=3.1.6
pyserial>=3.5
